pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "poolswitch"
version = "0.1.0"
description = "Production-grade PoolSwitch proxy with quota failover and multi-language SDKs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "OpenAI Codex" }]
keywords = ["proxy", "api", "rate-limit", "quota", "failover", "sdk"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Topic :: Internet :: Proxy Servers",
]
dependencies = [
  "aiosqlite>=0.20.0",
  "click>=8.1.7",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "prometheus-client>=0.22.0",
  "pydantic>=2.11.0",
  "pyyaml>=6.0.2",
  "redis>=5.2.1",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
poolswitch = "poolswitch.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["poolswitch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=poolswitch --cov=sdk-python/poolswitch_client --cov-branch --cov-report=term-missing --cov-fail-under=100"

